refactor to create CaSecret class, derived from the Secret class
